+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Allow replying with "conversation" visibility to author's own followers-only post.
 
+### Fixed
+
+- Don't change color of disabled "Publish" button on hover.
+
 ## [3.19.0] - 2025-03-09
 
 ### Changed